ESRDataImporterCamt54: trick jaxb into accepting also version 04 and 05
* also solve a bug related to multi trx-XML
* also add a public camt-54 test XML and do a simple test with it

1903: Support camt.054.001.04 explicitly 

Task-Url: https://github.com/metasfresh/metasfresh/issues/1903
