+++ de.metas.payment.esr/src/main/java/de/metas/payment/esr/dataimporter/IESRDataImporter.java
@@ -10,20 +10,24 @@ package de.metas.payment.esr.dataimporter;
  * it under the terms of the GNU General Public License as
  * published by the Free Software Foundation, either version 2 of the
  * License, or (at your option) any later version.
- * 
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public
- * License along with this program.  If not, see
+ * License along with this program. If not, see
  * <http://www.gnu.org/licenses/gpl-2.0.html>.
  * #L%
  */
 
-
 public interface IESRDataImporter
 {
+	/**
+	 * Import the data into our generic {@link ESRStatement} and {@link ESRTransaction} POJOs.
+	 *
+	 * @return
+	 */
 	ESRStatement importData();
 }
